+++ packages/adminui-vue/utils/replaceExtJSWidgets.js
@@ -4,7 +4,9 @@ const Vue = require('vue')
 const { Notification } = require('element-ui')
 const dialogs = require('../components/dialog/UDialog')
 const UNavbar = require('../components/navbar/UNavbar.vue').default
+const autoFormComponent = require('../components/AutoFormComponent.vue').default
 const { dialog, dialogInfo, dialogYesNo, dialogError } = dialogs
+const mountHelpers = require('./mountHelpers')
 
 function replaceExtJSDialogs () {
   // rename buttonText - > buttons, fn -> callback and call `dialog`
@@ -75,7 +77,20 @@ function replaceExtJSNavbar () {
   }).$mount(`#${id}`)
 }
 
+function replaceAutoForms () {
+  let params = this
+
+  if (mountHelpers.activateIfMounted(params)) return
+
+  let mountParams = {
+    FormComponent: autoFormComponent,
+    showFormParams: params
+  }
+  mountHelpers.mount(mountParams)
+}
+
 module.exports = {
   replaceExtJSDialogs,
-  replaceExtJSNavbar
+  replaceExtJSNavbar,
+  replaceAutoForms
 }